fix(popover): refactor imports and exports of InfoPopover
