+++ packages/components/src/checkbox/CheckboxGroup.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { AriaCheckboxGroupProps, useCheckboxGroup } from 'react-aria'
 import { useCheckboxGroupState } from 'react-stately'
-import { Label } from '../label'
+import { InfoPopoverProps, Label, LabelWrapper } from '../label'
 import { Text } from '../text'
 import { useLocalizedStringFormatter } from '../utils/intl'
 import { useSelectAll } from './useSelectAll'
@@ -10,8 +10,6 @@ import { CheckboxGroupContext } from './context'
 import { CheckboxGroupFieldError } from './CheckboxGroupFieldError'
 import messages from './intl/translations.json'
 import styles from './Checkbox.module.css'
-import { InfoPopoverProps } from '../label/InfoPopover'
-import { LabelWrapper } from '../label/LabelWrapper'
 
 export interface CheckboxGroupProps extends AriaCheckboxGroupProps {
   children: React.ReactNode
