+++ packages/components/src/date-picker/DateRangePicker.tsx
@@ -12,13 +12,11 @@ import { DatePickerInputField } from './DatePickerInputField'
 import { DatePickerPopover } from './DatePickerPopover'
 import { DateInput, DateInputDivider, DateSegment } from '../date-field'
 import { FieldError } from '../field-error'
-import { Label } from '../label'
+import { InfoPopoverProps, Label, LabelWrapper } from '../label'
 import { RangeCalendar } from '../calendar'
 import { Text } from '../text'
 import styles from './DatePicker.module.css'
 import { Size } from '../common/types'
-import { LabelWrapper } from '../label/LabelWrapper'
-import { InfoPopoverProps } from '../label/InfoPopover'
 
 export interface DateRangePickerProps
   extends AriaDateRangePickerProps<DateValue> {
