+++ packages/components/src/label/index.ts
@@ -1,1 +1,3 @@
 export * from './Label'
+export { type InfoPopoverProps } from './InfoPopover'
+export { LabelWrapper } from './LabelWrapper'
